8268118: Rename bytes_os_cpu.inline.hpp files to bytes_os_cpu.hpp
