+++ src/hotspot/cpu/s390/bytes_s390.hpp
@@ -51,7 +51,7 @@ class Bytes: AllStatic {
   static inline void put_native_u8(address p, u8 x) { *(u8*)p = x; }
 
   // The following header contains the implementations of swap_u2, swap_u4, and swap_u8.
-#include OS_CPU_HEADER_INLINE(bytes)
+#include OS_CPU_HEADER(bytes)
 
   // Efficient reading and writing of unaligned unsigned data in Java byte ordering (i.e. big-endian ordering)
   static inline u2   get_Java_u2(address p) { return get_native_u2(p); }
