+++ src/hotspot/cpu/x86/bytes_x86.hpp
@@ -122,6 +122,6 @@ class Bytes: AllStatic {
 };
 
 // The following header contains the implementations of swap_u2, swap_u4, and swap_u8[_base]
-#include OS_CPU_HEADER_INLINE(bytes)
+#include OS_CPU_HEADER(bytes)
 
 #endif // CPU_X86_BYTES_X86_HPP
