+++ src/hotspot/os_cpu/aix_ppc/bytes_aix_ppc.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,23 +22,11 @@
  *
  */
 
-#ifndef OS_CPU_WINDOWS_X86_BYTES_WINDOWS_X86_INLINE_HPP
-#define OS_CPU_WINDOWS_X86_BYTES_WINDOWS_X86_INLINE_HPP
+#ifndef OS_CPU_AIX_PPC_BYTES_AIX_PPC_HPP
+#define OS_CPU_AIX_PPC_BYTES_AIX_PPC_HPP
 
-#include <stdlib.h>
+#if defined(VM_LITTLE_ENDIAN)
+// Aix is not little endian.
+#endif // VM_LITTLE_ENDIAN
 
-// Efficient swapping of data bytes from Java byte
-// ordering to native byte ordering and vice versa.
-inline u2 Bytes::swap_u2(u2 x) {
-  return (u2) _byteswap_ushort((unsigned short) x);
-}
-
-inline u4 Bytes::swap_u4(u4 x) {
-  return (u4) _byteswap_ulong((unsigned long) x);
-}
-
-inline u8 Bytes::swap_u8(u8 x) {
-  return (u8) _byteswap_uint64((unsigned __int64) x);
-}
-
-#endif // OS_CPU_WINDOWS_X86_BYTES_WINDOWS_X86_INLINE_HPP
+#endif // OS_CPU_AIX_PPC_BYTES_AIX_PPC_HPP
