+++ src/hotspot/os_cpu/bsd_aarch64/bytes_bsd_aarch64.hpp
@@ -1,5 +1,7 @@
 /*
- * Copyright (c) 2008, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1999, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, Red Hat Inc. All rights reserved.
+ * Copyright (c) 2021, Azul Systems, Inc. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,26 +24,33 @@
  *
  */
 
-#ifndef OS_CPU_LINUX_ARM_BYTES_LINUX_ARM_INLINE_HPP
-#define OS_CPU_LINUX_ARM_BYTES_LINUX_ARM_INLINE_HPP
+#ifndef OS_CPU_BSD_AARCH64_BYTES_BSD_AARCH64_HPP
+#define OS_CPU_BSD_AARCH64_BYTES_BSD_AARCH64_HPP
 
-#include <byteswap.h>
+#ifdef __APPLE__
+#include <libkern/OSByteOrder.h>
+#endif
+
+#if defined(__APPLE__)
+#  define bswap_16(x) OSSwapInt16(x)
+#  define bswap_32(x) OSSwapInt32(x)
+#  define bswap_64(x) OSSwapInt64(x)
+#else
+#  error "Unimplemented"
+#endif
 
 // Efficient swapping of data bytes from Java byte
 // ordering to native byte ordering and vice versa.
-inline u2 Bytes::swap_u2(u2 x) {
-  // TODO: ARM - optimize
+inline u2   Bytes::swap_u2(u2 x) {
   return bswap_16(x);
 }
 
-inline u4 Bytes::swap_u4(u4 x) {
-  // TODO: ARM - optimize
+inline u4   Bytes::swap_u4(u4 x) {
   return bswap_32(x);
 }
 
 inline u8 Bytes::swap_u8(u8 x) {
-  // TODO: ARM - optimize
   return bswap_64(x);
 }
 
-#endif // OS_CPU_LINUX_ARM_BYTES_LINUX_ARM_INLINE_HPP
+#endif // OS_CPU_BSD_AARCH64_BYTES_BSD_AARCH64_HPP
