+++ src/hotspot/os_cpu/bsd_zero/bytes_bsd_zero.hpp
@@ -1,7 +1,5 @@
 /*
- * Copyright (c) 1999, 2021, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2014, Red Hat Inc. All rights reserved.
- * Copyright (c) 2021, Azul Systems, Inc. All rights reserved.
+ * Copyright (c) 2003, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,28 +22,41 @@
  *
  */
 
-#ifndef OS_CPU_BSD_AARCH64_BYTES_BSD_AARCH64_INLINE_HPP
-#define OS_CPU_BSD_AARCH64_BYTES_BSD_AARCH64_INLINE_HPP
-
-#ifdef __APPLE__
-#include <libkern/OSByteOrder.h>
-#endif
-
-#if defined(__APPLE__)
-#  define bswap_16(x) OSSwapInt16(x)
-#  define bswap_32(x) OSSwapInt32(x)
-#  define bswap_64(x) OSSwapInt64(x)
-#else
-#  error "Unimplemented"
-#endif
+#ifndef OS_CPU_BSD_ZERO_BYTES_BSD_ZERO_HPP
+#define OS_CPU_BSD_ZERO_BYTES_BSD_ZERO_HPP
 
 // Efficient swapping of data bytes from Java byte
 // ordering to native byte ordering and vice versa.
-inline u2   Bytes::swap_u2(u2 x) {
+
+#ifdef __APPLE__
+#  include <libkern/OSByteOrder.h>
+#else
+#  include <sys/endian.h>
+#endif
+
+#if defined(__APPLE__)
+#  define bswap_16(x)   OSSwapInt16(x)
+#  define bswap_32(x)   OSSwapInt32(x)
+#  define bswap_64(x)   OSSwapInt64(x)
+#elif defined(__OpenBSD__)
+#  define bswap_16(x)   swap16(x)
+#  define bswap_32(x)   swap32(x)
+#  define bswap_64(x)   swap64(x)
+#elif defined(__NetBSD__)
+#  define bswap_16(x)   bswap16(x)
+#  define bswap_32(x)   bswap32(x)
+#  define bswap_64(x)   bswap64(x)
+#else
+#  define bswap_16(x) __bswap16(x)
+#  define bswap_32(x) __bswap32(x)
+#  define bswap_64(x) __bswap64(x)
+#endif
+
+inline u2 Bytes::swap_u2(u2 x) {
   return bswap_16(x);
 }
 
-inline u4   Bytes::swap_u4(u4 x) {
+inline u4 Bytes::swap_u4(u4 x) {
   return bswap_32(x);
 }
 
@@ -53,4 +64,4 @@ inline u8 Bytes::swap_u8(u8 x) {
   return bswap_64(x);
 }
 
-#endif // OS_CPU_BSD_AARCH64_BYTES_BSD_AARCH64_INLINE_HPP
+#endif // OS_CPU_BSD_ZERO_BYTES_BSD_ZERO_HPP
