+++ src/hotspot/os_cpu/linux_aarch64/bytes_linux_aarch64.hpp
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2020, Microsoft Corporation. All rights reserved.
+ * Copyright (c) 1999, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, Red Hat Inc. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,25 +23,23 @@
  *
  */
 
-#ifndef OS_CPU_WINDOWS_AARCH64_BYTES_WINDOWS_AARCH64_INLINE_HPP
-#define OS_CPU_WINDOWS_AARCH64_BYTES_WINDOWS_AARCH64_INLINE_HPP
+#ifndef OS_CPU_LINUX_AARCH64_BYTES_LINUX_AARCH64_HPP
+#define OS_CPU_LINUX_AARCH64_BYTES_LINUX_AARCH64_HPP
 
-#include <stdlib.h>
+#include <byteswap.h>
 
 // Efficient swapping of data bytes from Java byte
 // ordering to native byte ordering and vice versa.
 inline u2   Bytes::swap_u2(u2 x) {
-  return _byteswap_ushort(x);
+  return bswap_16(x);
 }
 
 inline u4   Bytes::swap_u4(u4 x) {
-  return _byteswap_ulong(x);
+  return bswap_32(x);
 }
 
 inline u8 Bytes::swap_u8(u8 x) {
-  return _byteswap_uint64(x);
+  return bswap_64(x);
 }
 
-#pragma warning(default: 4035) // Enable warning 4035: no return value
-
-#endif // OS_CPU_WINDOWS_AARCH64_BYTES_WINDOWS_AARCH64_INLINE_HPP
+#endif // OS_CPU_LINUX_AARCH64_BYTES_LINUX_AARCH64_HPP
