+++ src/hotspot/os_cpu/linux_x86/bytes_linux_x86.hpp
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 1999, 2019, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2014, Red Hat Inc. All rights reserved.
+ * Copyright (c) 1999, 2020, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,23 +22,58 @@
  *
  */
 
-#ifndef OS_CPU_LINUX_AARCH64_BYTES_LINUX_AARCH64_INLINE_HPP
-#define OS_CPU_LINUX_AARCH64_BYTES_LINUX_AARCH64_INLINE_HPP
+#ifndef OS_CPU_LINUX_X86_BYTES_LINUX_X86_HPP
+#define OS_CPU_LINUX_X86_BYTES_LINUX_X86_HPP
 
 #include <byteswap.h>
 
 // Efficient swapping of data bytes from Java byte
 // ordering to native byte ordering and vice versa.
 inline u2   Bytes::swap_u2(u2 x) {
+#ifdef AMD64
   return bswap_16(x);
+#else
+  u2 ret;
+  __asm__ __volatile__ (
+    "movw %0, %%ax;"
+    "xchg %%al, %%ah;"
+    "movw %%ax, %0"
+    :"=r" (ret)      // output : register 0 => ret
+    :"0"  (x)        // input  : x => register 0
+    :"ax", "0"       // clobbered registers
+  );
+  return ret;
+#endif // AMD64
 }
 
 inline u4   Bytes::swap_u4(u4 x) {
+#ifdef AMD64
   return bswap_32(x);
+#else
+  u4 ret;
+  __asm__ __volatile__ (
+    "bswap %0"
+    :"=r" (ret)      // output : register 0 => ret
+    :"0"  (x)        // input  : x => register 0
+    :"0"             // clobbered register
+  );
+  return ret;
+#endif // AMD64
 }
 
+#ifdef AMD64
 inline u8 Bytes::swap_u8(u8 x) {
   return bswap_64(x);
 }
+#else
+// Helper function for swap_u8
+inline u8   Bytes::swap_u8_base(u4 x, u4 y) {
+  return (((u8)swap_u4(x))<<32) | swap_u4(y);
+}
 
-#endif // OS_CPU_LINUX_AARCH64_BYTES_LINUX_AARCH64_INLINE_HPP
+inline u8 Bytes::swap_u8(u8 x) {
+  return swap_u8_base(*(u4*)&x, *(((u4*)&x)+1));
+}
+#endif // !AMD64
+
+#endif // OS_CPU_LINUX_X86_BYTES_LINUX_X86_HPP
