+++ src/hotspot/os_cpu/windows_aarch64/bytes_windows_aarch64.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, Microsoft Corporation. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,11 +22,25 @@
  *
  */
 
-#ifndef OS_CPU_AIX_PPC_BYTES_AIX_PPC_INLINE_HPP
-#define OS_CPU_AIX_PPC_BYTES_AIX_PPC_INLINE_HPP
+#ifndef OS_CPU_WINDOWS_AARCH64_BYTES_WINDOWS_AARCH64_HPP
+#define OS_CPU_WINDOWS_AARCH64_BYTES_WINDOWS_AARCH64_HPP
 
-#if defined(VM_LITTLE_ENDIAN)
-// Aix is not little endian.
-#endif // VM_LITTLE_ENDIAN
+#include <stdlib.h>
 
-#endif // OS_CPU_AIX_PPC_BYTES_AIX_PPC_INLINE_HPP
+// Efficient swapping of data bytes from Java byte
+// ordering to native byte ordering and vice versa.
+inline u2   Bytes::swap_u2(u2 x) {
+  return _byteswap_ushort(x);
+}
+
+inline u4   Bytes::swap_u4(u4 x) {
+  return _byteswap_ulong(x);
+}
+
+inline u8 Bytes::swap_u8(u8 x) {
+  return _byteswap_uint64(x);
+}
+
+#pragma warning(default: 4035) // Enable warning 4035: no return value
+
+#endif // OS_CPU_WINDOWS_AARCH64_BYTES_WINDOWS_AARCH64_HPP
