+++ src/hotspot/os_cpu/windows_x86/bytes_windows_x86.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,46 +22,23 @@
  *
  */
 
-#ifndef OS_CPU_BSD_ZERO_BYTES_BSD_ZERO_INLINE_HPP
-#define OS_CPU_BSD_ZERO_BYTES_BSD_ZERO_INLINE_HPP
+#ifndef OS_CPU_WINDOWS_X86_BYTES_WINDOWS_X86_HPP
+#define OS_CPU_WINDOWS_X86_BYTES_WINDOWS_X86_HPP
+
+#include <stdlib.h>
 
 // Efficient swapping of data bytes from Java byte
 // ordering to native byte ordering and vice versa.
-
-#ifdef __APPLE__
-#  include <libkern/OSByteOrder.h>
-#else
-#  include <sys/endian.h>
-#endif
-
-#if defined(__APPLE__)
-#  define bswap_16(x)   OSSwapInt16(x)
-#  define bswap_32(x)   OSSwapInt32(x)
-#  define bswap_64(x)   OSSwapInt64(x)
-#elif defined(__OpenBSD__)
-#  define bswap_16(x)   swap16(x)
-#  define bswap_32(x)   swap32(x)
-#  define bswap_64(x)   swap64(x)
-#elif defined(__NetBSD__)
-#  define bswap_16(x)   bswap16(x)
-#  define bswap_32(x)   bswap32(x)
-#  define bswap_64(x)   bswap64(x)
-#else
-#  define bswap_16(x) __bswap16(x)
-#  define bswap_32(x) __bswap32(x)
-#  define bswap_64(x) __bswap64(x)
-#endif
-
 inline u2 Bytes::swap_u2(u2 x) {
-  return bswap_16(x);
+  return (u2) _byteswap_ushort((unsigned short) x);
 }
 
 inline u4 Bytes::swap_u4(u4 x) {
-  return bswap_32(x);
+  return (u4) _byteswap_ulong((unsigned long) x);
 }
 
 inline u8 Bytes::swap_u8(u8 x) {
-  return bswap_64(x);
+  return (u8) _byteswap_uint64((unsigned __int64) x);
 }
 
-#endif // OS_CPU_BSD_ZERO_BYTES_BSD_ZERO_INLINE_HPP
+#endif // OS_CPU_WINDOWS_X86_BYTES_WINDOWS_X86_HPP
